# specs/openapi/master-data.v1.yaml
openapi: 3.1.0
info:
  title: HIMS Master Data API
  version: 1.0.0
  description: 'HIMS Master Data catalog API — platform reference data (modules first; more surfaces phased in). Base path
    `/api/v1/master-data`. DELETE on modules is a recursive soft-delete (`is_deleted = true`). Gateway-level authentication
    (API keys / JWT) is expected before production; this spec does not model per-route security yet.

    '
servers:
- url: /api/v1/master-data
paths:
...
  /visitpad/rx-columns:
    get:
      tags:
      - Visitpad — Rx columns
      summary: List Rx columns
      operationId: get_visitpad_rx_columns_api_v1_master_data_visitpad_rx_columns_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: section
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/VisitpadRxColumnSection'
          - type: 'null'
          title: Section
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadRxColumnListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Visitpad — Rx columns
      summary: Create Rx column
      operationId: post_visitpad_rx_column_api_v1_master_data_visitpad_rx_columns_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitpadRxColumnCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadRxColumnSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /visitpad/rx-columns/{rx_column_id}:
    get:
      tags:
      - Visitpad — Rx columns
      summary: Get Rx column
      operationId: get_visitpad_rx_column_api_v1_master_data_visitpad_rx_columns__rx_column_id__get
      parameters:
      - name: rx_column_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Rx Column Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadRxColumnSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Visitpad — Rx columns
      summary: Update Rx column
      operationId: patch_visitpad_rx_column_api_v1_master_data_visitpad_rx_columns__rx_column_id__patch
      parameters:
      - name: rx_column_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Rx Column Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitpadRxColumnUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadRxColumnSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Visitpad — Rx columns
      summary: Soft-delete Rx column
      operationId: delete_visitpad_rx_column_api_v1_master_data_visitpad_rx_columns__rx_column_id__delete
      parameters:
      - name: rx_column_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Rx Column Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadRxColumnSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /visitpad/procedures:
    get:
      tags:
      - Visitpad — Procedures
      summary: List procedures
      operationId: get_procedures_api_v1_master_data_visitpad_procedures_get
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Search
      - name: category
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/VisitpadProcedureCategory'
          - type: 'null'
          title: Category
      - name: billing_category
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/VisitpadProcedureBillingCategory'
          - type: 'null'
          title: Billing Category
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadProcedureListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Visitpad — Procedures
      summary: Create procedure
      operationId: post_procedure_api_v1_master_data_visitpad_procedures_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitpadProcedureCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadProcedureSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /visitpad/procedures/{procedure_id}:
    get:
      tags:
      - Visitpad — Procedures
      summary: Get procedure
      operationId: get_procedure_api_v1_master_data_visitpad_procedures__procedure_id__get
      parameters:
      - name: procedure_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Procedure Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadProcedureSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Visitpad — Procedures
      summary: Update procedure
      operationId: patch_procedure_api_v1_master_data_visitpad_procedures__procedure_id__patch
      parameters:
      - name: procedure_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Procedure Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VisitpadProcedureUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadProcedureSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - Visitpad — Procedures
      summary: Soft-delete procedure
      operationId: delete_procedure_api_v1_master_data_visitpad_procedures__procedure_id__delete
      parameters:
      - name: procedure_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Procedure Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VisitpadProcedureSingleResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VisitpadProcedureBillingCategory:
      type: string
      enum:
      - professional
      - facility
      - ancillary
      - bundled
      - other
      title: VisitpadProcedureBillingCategory
    VisitpadProcedureCategory:
      type: string
      enum:
      - diagnostic
      - therapeutic
      - surgical
      - ancillary
      - other
      title: VisitpadProcedureCategory
    VisitpadProcedureCreate:
      properties:
        cpt_code:
          type: string
          maxLength: 16
          minLength: 1
          title: Cpt Code
        official_descriptor:
          type: string
          maxLength: 512
          minLength: 1
          title: Official Descriptor
        display_name:
          type: string
          maxLength: 512
          minLength: 1
          title: Display Name
        category:
          $ref: '#/components/schemas/VisitpadProcedureCategory'
        billing_category:
          $ref: '#/components/schemas/VisitpadProcedureBillingCategory'
        duration_minutes:
          type: integer
          maximum: 1440.0
          minimum: 0.0
          title: Duration Minutes
        requires_consent:
          type: boolean
          title: Requires Consent
          default: false
        type_modality:
          anyOf:
          - type: string
            maxLength: 128
          - type: 'null'
          title: Type Modality
        display_order:
          type: integer
          title: Display Order
          default: 0
        is_active:
          type: boolean
          title: Is Active
          default: true
        snomed_code:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Snomed Code
      additionalProperties: false
      type: object
      required:
      - cpt_code
      - official_descriptor
      - display_name
      - category
      - billing_category
      - duration_minutes
      title: VisitpadProcedureCreate
    VisitpadProcedureListResponse:
      properties:
        data:
          items:
            $ref: '#/components/schemas/VisitpadProcedureResponse'
          type: array
          title: Data
        total:
          type: integer
          title: Total
      type: object
      required:
      - data
      - total
      title: VisitpadProcedureListResponse
    VisitpadProcedureResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        tenant_id:
          type: string
          format: uuid
          title: Tenant Id
        cpt_code:
          type: string
          title: Cpt Code
        official_descriptor:
          type: string
          title: Official Descriptor
        display_name:
          type: string
          title: Display Name
        category:
          $ref: '#/components/schemas/VisitpadProcedureCategory'
        billing_category:
          $ref: '#/components/schemas/VisitpadProcedureBillingCategory'
        duration_minutes:
          type: integer
          title: Duration Minutes
        requires_consent:
          type: boolean
          title: Requires Consent
        type_modality:
          anyOf:
          - type: string
          - type: 'null'
          title: Type Modality
        display_order:
          type: integer
          title: Display Order
        is_active:
          type: boolean
          title: Is Active
        is_deleted:
          type: boolean
          title: Is Deleted
        snomed_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Snomed Code
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - tenant_id
      - cpt_code
      - official_descriptor
      - display_name
      - category
      - billing_category
      - duration_minutes
      - requires_consent
      - display_order
      - is_active
      - is_deleted
      - created_at
      - updated_at
      title: VisitpadProcedureResponse
    VisitpadProcedureSingleResponse:
      properties:
        data:
          $ref: '#/components/schemas/VisitpadProcedureResponse'
      type: object
      required:
      - data
      title: VisitpadProcedureSingleResponse
    VisitpadProcedureUpdate:
      properties:
        cpt_code:
          anyOf:
          - type: string
            maxLength: 16
            minLength: 1
          - type: 'null'
          title: Cpt Code
        official_descriptor:
          anyOf:
          - type: string
            maxLength: 512
            minLength: 1
          - type: 'null'
          title: Official Descriptor
        display_name:
          anyOf:
          - type: string
            maxLength: 512
            minLength: 1
          - type: 'null'
          title: Display Name
        category:
          anyOf:
          - $ref: '#/components/schemas/VisitpadProcedureCategory'
          - type: 'null'
        billing_category:
          anyOf:
          - $ref: '#/components/schemas/VisitpadProcedureBillingCategory'
          - type: 'null'
        duration_minutes:
          anyOf:
          - type: integer
            maximum: 1440.0
            minimum: 0.0
          - type: 'null'
          title: Duration Minutes
        requires_consent:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Requires Consent
        type_modality:
          anyOf:
          - type: string
            maxLength: 128
          - type: 'null'
          title: Type Modality
        display_order:
          anyOf:
          - type: integer
          - type: 'null'
          title: Display Order
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
        snomed_code:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Snomed Code
        is_deleted:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Deleted
      additionalProperties: false
      type: object
      title: VisitpadProcedureUpdate
...
    VisitpadRxColumnCreate:
      properties:
        section:
          $ref: '#/components/schemas/VisitpadRxColumnSection'
        display_name:
          type: string
          maxLength: 256
          minLength: 1
          title: Display Name
        code:
          type: string
          maxLength: 64
          minLength: 1
          title: Code
        extra_unit:
          anyOf:
          - type: string
            maxLength: 128
          - type: 'null'
          title: Extra Unit
        display_order:
          type: integer
          title: Display Order
          default: 0
        is_active:
          type: boolean
          title: Is Active
          default: true
      additionalProperties: false
      type: object
      required:
      - section
      - display_name
      - code
      title: VisitpadRxColumnCreate
    VisitpadRxColumnListResponse:
      properties:
        data:
          items:
            $ref: '#/components/schemas/VisitpadRxColumnResponse'
          type: array
          title: Data
        total:
          type: integer
          title: Total
      type: object
      required:
      - data
      - total
      title: VisitpadRxColumnListResponse
    VisitpadRxColumnResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        tenant_id:
          type: string
          format: uuid
          title: Tenant Id
        section:
          $ref: '#/components/schemas/VisitpadRxColumnSection'
        display_name:
          type: string
          title: Display Name
        code:
          type: string
          title: Code
        extra_unit:
          anyOf:
          - type: string
          - type: 'null'
          title: Extra Unit
        display_order:
          type: integer
          title: Display Order
        is_active:
          type: boolean
          title: Is Active
        is_deleted:
          type: boolean
          title: Is Deleted
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - tenant_id
      - section
      - display_name
      - code
      - display_order
      - is_active
      - is_deleted
      - created_at
      - updated_at
      title: VisitpadRxColumnResponse
    VisitpadRxColumnSection:
      type: string
      enum:
      - medication_type
      - frequency
      - unit
      - diet_type
      - method_strength
      - route
      - time_of_administration
      title: VisitpadRxColumnSection
    VisitpadRxColumnSingleResponse:
      properties:
        data:
          $ref: '#/components/schemas/VisitpadRxColumnResponse'
      type: object
      required:
      - data
      title: VisitpadRxColumnSingleResponse
    VisitpadRxColumnUpdate:
      properties:
        display_name:
          anyOf:
          - type: string
            maxLength: 256
            minLength: 1
          - type: 'null'
          title: Display Name
        code:
          anyOf:
          - type: string
            maxLength: 64
            minLength: 1
          - type: 'null'
          title: Code
        extra_unit:
          anyOf:
          - type: string
            maxLength: 128
          - type: 'null'
          title: Extra Unit
        display_order:
          anyOf:
          - type: integer
          - type: 'null'
          title: Display Order
        is_active:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Active
        is_deleted:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Deleted
          description: Set false to restore a soft-deleted row.
      additionalProperties: false
      type: object
      title: VisitpadRxColumnUpdate
...